More flex work.

+++ crates/css/modules/core/src/8_box_model/part_8_3_1_collapsing_margins.rs
@@ -527,5 +527,5 @@ pub fn compute_root_y_after_top_collapse(
     layouter
         .rects
         .get(&root)
-        .map_or(metrics.margin_top, |rect| rect.y)
+        .map_or(metrics.margin_top, |rect| rect.y as i32)
 }
